Specify when a message is a send
This expands the bridge -> client protocol to add a `send` key. It will
be `true` if the message is the result of a send, `false` if the result
of a publish.

Depends on https://github.com/eclipse/vert.x/commit/70b2b541e780d09e6c83cb434a2e977d6e3b550c

Fixes #20

Signed-off-by: Toby Crawley <toby@tcrawley.org>