Only deliver sends to one handler in the node client
This uses the new `send` key to know if a message should be delivered to
one or all handlers. This also fixes the behavior when an error message
is received - error messages that are directed at addresses are of type
"message", not "err". "err" messages a general, and have no address, so
should only go to the global error handler.

Related to #20

Signed-off-by: Toby Crawley <toby@tcrawley.org>